Added mouse to world space input Block removal with mouse input currently not working

+++ src/world/Block.cpp
@@ -2,5 +2,5 @@
 
 using namespace world;
 
-Block::Block(const MATERIAL type, const glm::ivec2 position,const Layer layer,const uint8_t luminance) : type(type),position(position),layer(layer),luminance(luminance)
+Block::Block(const MATERIAL type, const glm::vec2 position,const Layer layer,const uint8_t luminance) : type(type),position(position),layer(layer),luminance(luminance)
 {}
